[修复]修复删除文件的逻辑

+++ delete.js
@@ -13,21 +13,29 @@ const path = require('path');
   function check() {
     if (type !== 'local' && type !== 'remote') throw ('参数错误！');
     // 检查要删除的文件是否存在
-    const exists = fs.existsSync(path.resolve(__dirname, `${htmlPath}${name}.html`));
-    if (!exists) throw ('无法找到要删除的文件！');
     if (type === 'local') {
+      const existsLocal = fs.existsSync(path.resolve(__dirname, `${htmlPath}${name}.html`));
+      if (!existsLocal) throw ('无法找到要删除的文件！');
       delLocalFile();
     } else if (type === 'remote') {
+      const existsRemote = fs.existsSync(path.resolve(__dirname, `${htmlRemotePath}${name}.html`));
+      if (!existsRemote) throw ('无法找到要删除的文件！');
       delRemoteFile();
     }
   }
   check();
   // 删除本地开发环境的源代码
   function delLocalFile() {
+    // 删除之前赢检查是否存在
+    const existsCss = fs.existsSync(path.resolve(__dirname, `${staticPath}css/${name}.scss`));
+    const existsJs = fs.existsSync(path.resolve(__dirname, `${staticPath}js/${name}.js`));
     fs.unlinkSync(path.resolve(__dirname, `${htmlPath}${name}.html`)); // 删除html文件
-    fs.unlinkSync(path.resolve(__dirname, `${staticPath}css/${name}.scss`)); // 删除scss文件
-    fs.unlinkSync(path.resolve(__dirname, `${staticPath}js/${name}.js`)); // 删除js文件
-
+    if (existsCss) {
+      fs.unlinkSync(path.resolve(__dirname, `${staticPath}css/${name}.scss`)); // 删除scss文件
+    }
+    if (existsJs) {
+      fs.unlinkSync(path.resolve(__dirname, `${staticPath}js/${name}.js`)); // 删除js文件
+    }
     // 删除image文件
     let imagesName = fs.readdirSync(path.resolve(__dirname, `${staticPath}images/`));
     const reg = new RegExp(`${name}_`);
@@ -43,8 +51,8 @@ const path = require('path');
   // 删除远端生产环境的打包代码
   function delRemoteFile() {
     // 删除之前应检查是否存在
-    let existsCss = fs.existsSync(path.resolve(__dirname, `${staticRemotePath}css/${name}.css`));
-    let existsJs = fs.existsSync(path.resolve(__dirname, `${staticRemotePath}js/${name}.js`));
+    const existsCss = fs.existsSync(path.resolve(__dirname, `${staticRemotePath}css/${name}.css`));
+    const existsJs = fs.existsSync(path.resolve(__dirname, `${staticRemotePath}js/${name}.js`));
     fs.unlinkSync(path.resolve(__dirname, `${htmlRemotePath}${name}.html`)); // 删除html文件
     if (existsCss) {
       fs.unlinkSync(path.resolve(__dirname, `${staticRemotePath}css/${name}.css`)); // 删除css文件
